requirements.txt
Django==3.1.4
fontawesome-free==5.15.1
django-crispy-forms==1.10.0